refactor: move scan.execute() metrics test to scan.rs, reflow comments to 100-col width
